box: move func_cache to a separate file
I'm going to extend func cache API so it should be separated.
A couple of comments added.
No logical changes.

NO_DOC=refactoring
NO_CHANGELOG=refactoring
NO_TEST=refactoring
