+++ src/box/schema.cc
@@ -29,7 +29,6 @@
  * SUCH DAMAGE.
  */
 #include "schema.h"
-#include "func.h"
 #include "sequence.h"
 #include "tuple.h"
 #include "assoc.h"
@@ -59,8 +58,6 @@
 /** All existing spaces. */
 static struct mh_i32ptr_t *spaces;
 static struct mh_strnptr_t *spaces_by_name;
-static struct mh_i32ptr_t *funcs;
-static struct mh_strnptr_t *funcs_by_name;
 static struct mh_i32ptr_t *sequences;
 /** Public change counter. On its update clients need to fetch
  *  new space data from the instance. */
@@ -383,8 +380,7 @@ schema_init(void)
 	/* Initialize the space cache. */
 	spaces = mh_i32ptr_new();
 	spaces_by_name = mh_strnptr_new();
-	funcs = mh_i32ptr_new();
-	funcs_by_name = mh_strnptr_new();
+	func_cache_init();
 	sequences = mh_i32ptr_new();
 	/*
 	 * Create surrogate space objects for the mandatory system
@@ -537,16 +533,8 @@ schema_free(void)
 	}
 	mh_i32ptr_delete(spaces);
 	mh_strnptr_delete(spaces_by_name);
-	while (mh_size(funcs) > 0) {
-		mh_int_t i = mh_first(funcs);
+	func_cache_destroy();
 
-		struct func *func = ((struct func *)
-				     mh_i32ptr_node(funcs, i)->val);
-		func_cache_delete(func->def->fid);
-		func_delete(func);
-	}
-	mh_i32ptr_delete(funcs);
-	mh_strnptr_delete(funcs_by_name);
 	while (mh_size(sequences) > 0) {
 		mh_int_t i = mh_first(sequences);
 
@@ -555,53 +543,6 @@ schema_free(void)
 		sequence_cache_delete(seq->def->id);
 	}
 	mh_i32ptr_delete(sequences);
-}
-
-void
-func_cache_insert(struct func *func)
-{
-	assert(func_by_id(func->def->fid) == NULL);
-	assert(func_by_name(func->def->name, strlen(func->def->name)) == NULL);
-	const struct mh_i32ptr_node_t node = { func->def->fid, func };
-	mh_i32ptr_put(funcs, &node, NULL, NULL);
-	size_t def_name_len = strlen(func->def->name);
-	uint32_t name_hash = mh_strn_hash(func->def->name, def_name_len);
-	const struct mh_strnptr_node_t strnode = {
-		func->def->name, def_name_len, name_hash, func };
-	mh_strnptr_put(funcs_by_name, &strnode, NULL, NULL);
-}
-
-void
-func_cache_delete(uint32_t fid)
-{
-	mh_int_t k = mh_i32ptr_find(funcs, fid, NULL);
-	if (k == mh_end(funcs))
-		return;
-	struct func *func = (struct func *)
-		mh_i32ptr_node(funcs, k)->val;
-	mh_i32ptr_del(funcs, k, NULL);
-	k = mh_strnptr_find_inp(funcs_by_name, func->def->name,
-				strlen(func->def->name));
-	if (k != mh_end(funcs))
-		mh_strnptr_del(funcs_by_name, k, NULL);
-}
-
-struct func *
-func_by_id(uint32_t fid)
-{
-	mh_int_t func = mh_i32ptr_find(funcs, fid, NULL);
-	if (func == mh_end(funcs))
-		return NULL;
-	return (struct func *) mh_i32ptr_node(funcs, func)->val;
-}
-
-struct func *
-func_by_name(const char *name, uint32_t name_len)
-{
-	mh_int_t func = mh_strnptr_find_inp(funcs_by_name, name, name_len);
-	if (func == mh_end(funcs_by_name))
-		return NULL;
-	return (struct func *) mh_strnptr_node(funcs_by_name, func)->val;
 }
 
 int
